Fixes for menu bar Add more animations in main menu
