+++ Assets/Animations/SeedlingSelector.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 78ab53b8b8e87254d865652d2be24d68
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
