+++ Assets/Animations/SlideMenuBar.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9823a0457f6b6414d8b8a98f184cc8f6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
