+++ Assets/Animations/SlideSeedlingSelector.anim.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f688c7582882b5b4a8050af6145668a3
+guid: af0a12b43fab05a4db3357e48d611f5c
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 0
