+++ Assets/SlideSelector.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c4cf1077032e7d346a9cc1c5573a7207
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
