+++ Assets/Sprites/UI/SettingsIcon.png.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5666e65d31e376b49917d9a48ba455fd
+guid: b5ee14c3d21771f48975840b1526ee0a
 TextureImporter:
   internalIDToNameTable: []
   externalObjects: {}
@@ -46,8 +46,8 @@ TextureImporter:
   spriteMeshType: 1
   alignment: 0
   spritePivot: {x: 0.5, y: 0.5}
-  spritePixelsToUnits: 10
-  spriteBorder: {x: 0, y: 3, z: 2, w: 3}
+  spritePixelsToUnits: 100
+  spriteBorder: {x: 0, y: 0, z: 0, w: 0}
   spriteGenerateFallbackPhysicsShape: 1
   alphaUsage: 1
   alphaIsTransparency: 1
